+++ kkambo/chap02/ex/exQ9.java
@@ -1,9 +1,8 @@
-package kkambo.chap02;
+package kkambo.chap02.ex;
 
 import java.util.Scanner;
 
-public class exQ8 {
-
+public class exQ9 {
     static int[][] mdays = {
             {31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31},//평년
             {31, 29, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31},//윤년
@@ -14,13 +13,16 @@ public class exQ8 {
         return (year % 4 == 0 && year % 100 !=0 || year % 400 == 0) ? 1: 0;
     }
 
-    static int dayOfYear(int y, int m, int d){
+    static int leftDayOfYear(int y, int m, int d){
+        int days = mdays[isLeap(y)][m-1] - d;
 
-        do{
-            d += mdays[isLeap(y)][--m];
-        }while(m > 1);
+        System.out.println("days = " + days);
 
-        return d;
+        while(m < 12) {
+            days += mdays[isLeap(y)][m++];
+        }
+
+        return days;
     }
 
     public static void main(String[] args) {
@@ -33,7 +35,7 @@ public class exQ8 {
         System.out.print("일 : ");
         int day = stdIn.nextInt();
 
-        System.out.printf("그 해 %d 일째입니다.", dayOfYear(year, month, day));
+        System.out.printf("그 해 %d 일째입니다.", leftDayOfYear(year, month, day));
     }
 
 }
